refactor: Move get_export_extension function
