+++ quadra/utils/tests/helpers.py
@@ -7,6 +7,7 @@ from hydra import compose, initialize_config_module
 from hydra.core.hydra_config import HydraConfig
 
 from quadra.main import main
+from quadra.utils.export import get_export_extension
 
 
 # taken from hydra unit tests
@@ -28,27 +29,6 @@ def execute_quadra_experiment(overrides: List[str], experiment_path: Path) -> No
         HydraConfig.instance().set_config(cfg)
 
         main(cfg)
-
-
-def get_export_extension(export_type: str) -> str:
-    """Get the extension of the exported model.
-
-    Args:
-        export_type: The type of the exported model.
-
-    Returns:
-        The extension of the exported model.
-    """
-    if export_type == "onnx":
-        extension = "onnx"
-    elif export_type == "torchscript":
-        extension = "pt"
-    elif export_type == "pytorch":
-        extension = "pth"
-    else:
-        raise ValueError(f"Unsupported export type {export_type}")
-
-    return extension
 
 
 def check_deployment_model(export_type: str):
