+++ tests/tasks/test_anomaly.py
@@ -6,8 +6,9 @@ from typing import List
 
 import pytest
 
+from quadra.utils.export import get_export_extension
 from quadra.utils.tests.fixtures import base_anomaly_dataset
-from quadra.utils.tests.helpers import check_deployment_model, execute_quadra_experiment, get_export_extension
+from quadra.utils.tests.helpers import check_deployment_model, execute_quadra_experiment
 
 try:
     import onnx  # noqa
